--- packages/react/src/experimental/OneLane/index.tsx
@@ -1,1 +1,0 @@
-export * from "./OneLane"
